pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfscalpel"
version = "0.1.0"
description = "Forensic-grade PDF analysis and CTF toolkit"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "PDFScalpel Contributors"}
]
keywords = ["pdf", "forensics", "ctf", "security", "analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Utilities",
]

dependencies = [
    "pikepdf>=8.0.0",
    "pdfplumber>=0.10.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tomli>=2.0.0; python_version<'3.11'",
]

[project.optional-dependencies]
full = [
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pycryptodome>=3.18.0",
    "ocrmypdf>=15.0.0",
    "graphviz>=0.20.0",
    "python-magic-bin>=0.4.14; platform_system=='Windows'",
    "python-magic>=0.4.27; platform_system!='Windows'",
    "pypdf>=3.17.0",
    "httpx>=0.27.0",
    "browser-cookie3>=0.19.0",
    "reportlab>=4.0.0",
]

dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
pdfscalpel = "pdfscalpel.cli.main:cli"

[project.urls]
Homepage = "https://github.com/pdfscalpel/pdfscalpel"
Documentation = "https://github.com/pdfscalpel/pdfscalpel#readme"
Repository = "https://github.com/pdfscalpel/pdfscalpel"
Issues = "https://github.com/pdfscalpel/pdfscalpel/issues"

[tool.setuptools]
packages = [
    "pdfscalpel",
    "pdfscalpel.analyze",
    "pdfscalpel.cli",
    "pdfscalpel.core",
    "pdfscalpel.extract",
    "pdfscalpel.generate",
    "pdfscalpel.integrations",
    "pdfscalpel.mutate",
    "pdfscalpel.plugins",
    "pdfscalpel.plugins.examples",
    "pdfscalpel.solve",
]

[tool.setuptools.package-data]
pdfscalpel = ["py.typed", "*.toml.example"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=pdfscalpel --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PT"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
